Fix: open the browser on Linux
Detects if the OS is Linux and uses xdg-open to open the generated link.
xdg-open is the Linux equivalent of Mac's open command.

"xdg-open opens a file or URL in the user's preferred application. If a URL is provided the URL will be opened in the user's preferred web browser. If a file is provided the file will be opened in the preferred application for files of that type. xdg-open supports file, ftp, http and https URLs."

https://linux.die.net/man/1/xdg-open